refactor: fix mock for registration member
