+++ packages/web/src/common/components/RegisterMemberTable.tsx
@@ -9,7 +9,6 @@ import { ApiReg019ResponseOk } from "@sparcs-clubs/interface/api/registration/en
 
 import Table from "@sparcs-clubs/web/common/components/Table";
 import Tag from "@sparcs-clubs/web/common/components/Tag";
-import mockupRegistrationMember from "@sparcs-clubs/web/features/executive/register-member/_mock/mockMemberRegistration";
 import {
   getTagColorFromClubType,
   getTagColorFromDivision,
@@ -20,8 +19,7 @@ interface RegisterMemberTableProps {
   registerMemberList: ApiReg019ResponseOk;
 }
 
-const columnHelper =
-  createColumnHelper<(typeof mockupRegistrationMember)["items"][number]>();
+const columnHelper = createColumnHelper<ApiReg019ResponseOk["items"][number]>();
 
 const columns = [
   columnHelper.accessor("clubTypeEnumId", {
